Move MessageText and MessageGetFiles traits to new util module
