+++ lib/src/lib.rs
@@ -10,8 +10,8 @@ mod stream;
 
 pub mod connector;
 pub mod prelude;
-pub mod traits;
 pub mod types;
+pub mod util;
 
 pub use self::api::Api;
 pub use self::errors::{Error, ErrorKind};
