+++ lib/src/prelude.rs
@@ -22,4 +22,4 @@ pub use telegram_bot_raw::{CanReplySendVenue, CanSendVenue};
 pub use telegram_bot_raw::{CanUnbanChatMemberForChat, CanUnbanChatMemberForUser};
 pub use telegram_bot_raw::{ToReplyRequest, ToRequest};
 
-pub use crate::traits::{MessageText, MessageGetFiles};
+pub use crate::util::messages::{MessageText, MessageGetFiles};
